Move game-state.mjs into gamelib folder
