+++ services/gateway/gateway-server.mjs
@@ -26,7 +26,7 @@ import AuthServer from '../auth/AuthServer.mjs';
 
 /* Import all utility classes */
 import GameMap from '../../common/maplib/map.mjs';
-import GameState from '../../common/maplib/game-state.mjs';
+import GameState from '../../common/gamelib/game-state.mjs';
 import ExtensionManager from '../../common/extlib/extension-manager.mjs';
 
 async function mainServer() {
