+++ sites/game-client/client.ejs
@@ -55,7 +55,7 @@ absolute URL.
     <script type="module">
       import GameMap from '/static/common/maplib/map.mjs';
       import GraphicAsset from '/static/common/maplib/graphic-asset.mjs';
-      import GameState from '/static/common/maplib/game-state.mjs';
+      import GameState from '/static/common/gamelib/game-state.mjs';
       import GameClient from '/static/sites/game-client/game-client.mjs';
       import MapRenderer from '/static/sites/game-client/map-renderer.mjs';
       import InputManager from '/static/sites/game-client/input-manager.mjs';
